update config file

+++ dataset.py
@@ -58,8 +58,8 @@ class FashionDataset(Dataset):
 
         sequence_features = [self.articles_dict[article] for article in self.samples['sequence'][index]]
         customer_features = self.customers_dict[self.samples['customer_id'][index]]
-        sequence_features = torch.tensor(sequence_features, device=device)
-        customer_features = torch.tensor(customer_features, device=device)
+        sequence_features = torch.tensor(sequence_features, dtype=torch.float, device=device)
+        customer_features = torch.tensor(customer_features, dtype=torch.float, device=device)
         return {'sequence_features': sequence_features, 'customer_features': customer_features, 'target': torch.tensor(target, device=device)}
